+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## [2.0.11] - 2025-04-24
+- Show error ID instead of code. This is handy when you want to suppress an error. 
+
+## [2.0.10] - 2025-04-24
+- Fix file not being founf via compile commands for parsing.
+- Fix defines not being parsed correctly from compile commands.
+
+## [2.0.9] - 2025-04-21
+- Parse compile_commands.json for defines and include directories. Cppcheck is no longer called with the project flag. This allows adding include directories and in the future even modifying them. 
+
+## [2.0.8] - 2025-03-01
+- Correct source of problems showing Cppcheck-Turbo in problems tab
+
 ## [2.0.7] - 2025-03-01
 - Fix bug where cppchecking would stay visible if no .cppcheck-config is found.